+++ src/migrations/M190319185130Cms.php
@@ -2,30 +2,30 @@
 
 namespace davidhirtz\yii2\cms\migrations;
 
+use davidhirtz\yii2\cms\migrations\traits\I18nTablesTrait;
 use davidhirtz\yii2\cms\models\Entry;
 use davidhirtz\yii2\cms\models\Section;
 use davidhirtz\yii2\cms\Module;
-use davidhirtz\yii2\cms\modules\ModuleTrait;
 use davidhirtz\yii2\skeleton\db\traits\MigrationTrait;
 use davidhirtz\yii2\skeleton\models\User;
 use Yii;
 use yii\db\Migration;
 
+/**
+ * @noinspection PhpUnused
+ */
+
 class M190319185130Cms extends Migration
 {
     use MigrationTrait;
-    use ModuleTrait;
+    use I18nTablesTrait;
 
     public function safeUp(): void
     {
-        $schema = $this->getDb()->getSchema();
+        $this->i18nTablesCallback(function () {
+            $schema = $this->getDb()->getSchema();
 
-        foreach ($this->getLanguages() as $language) {
-            if ($language) {
-                Yii::$app->language = $language;
-            }
-
-            // Entry.
+            // Entry
             $this->createTable(Entry::tableName(), [
                 'id' => $this->primaryKey()->unsigned(),
                 'status' => $this->tinyInteger(1)->unsigned()->notNull()->defaultValue(Entry::STATUS_ENABLED),
@@ -54,7 +54,7 @@ class M190319185130Cms extends Migration
             $tableName = $schema->getRawTableName(Entry::tableName());
             $this->addForeignKey($tableName . '_updated_by_ibfk', Entry::tableName(), 'updated_by_user_id', User::tableName(), 'id', 'SET NULL');
 
-            // Section.
+            // Section
             $this->createTable(Section::tableName(), [
                 'id' => $this->primaryKey()->unsigned(),
                 'status' => $this->tinyInteger(1)->unsigned()->notNull()->defaultValue(Section::STATUS_ENABLED),
@@ -78,7 +78,7 @@ class M190319185130Cms extends Migration
             $tableName = $schema->getRawTableName(Section::tableName());
             $this->addForeignKey($tableName . '_entry_id_ibfk', Section::tableName(), 'entry_id', Entry::tableName(), 'id', 'CASCADE');
             $this->addForeignKey($tableName . '_updated_by_ibfk', Section::tableName(), 'updated_by_user_id', User::tableName(), 'id', 'SET NULL');
-        }
+        });
 
         $auth = Yii::$app->getAuthManager();
         $admin = $auth->getRole(User::AUTH_ROLE_ADMIN);
@@ -91,21 +91,14 @@ class M190319185130Cms extends Migration
 
     public function safeDown(): void
     {
-        foreach ($this->getLanguages() as $language) {
-            Yii::$app->language = $language;
-
+        $this->i18nTablesCallback(function () {
             $this->dropTable(Section::tableName());
             $this->dropTable(Entry::tableName());
-        }
+        });
 
         $auth = Yii::$app->getAuthManager();
         $this->delete($auth->itemTable, ['name' => Module::AUTH_ROLE_AUTHOR]);
 
         $auth->invalidateCache();
-    }
-
-    private function getLanguages(): array
-    {
-        return static::getModule()->enableI18nTables ? Yii::$app->getI18n()->getLanguages() : [Yii::$app->language];
     }
 }