+++ src/models/queries/EntryQuery.php
@@ -103,7 +103,7 @@ class EntryQuery extends I18nActiveQuery
     public function whereSection(Section $section, bool $eagerLoading = true, string $joinType = 'INNER JOIN'): static
     {
         $tableName = SectionEntry::tableName();
-        $onCondition = fn(ActiveQuery $query) => $query->onCondition(["$tableName.[[section_id]]" => $section->id]);
+        $onCondition = fn (ActiveQuery $query) => $query->onCondition(["$tableName.[[section_id]]" => $section->id]);
 
         if ($eagerLoading && $joinType === 'INNER JOIN') {
             $this->orderBy(["$tableName.[[position]]" => SORT_ASC]);