+++ src/modules/admin/widgets/forms/traits/FooterFieldTrait.php
@@ -9,6 +9,7 @@ use yii\widgets\ActiveField;
 /**
  * @property FooterAttributeTrait $model
  * @mixin EntryActiveForm
+ * @noinspection PhpUnused
  */
 trait FooterFieldTrait
 {